fix(lib): added handling of add_offset and scale_factor. refactored code
