+++ src/lib/data/coordinateVariables.ts
@@ -1,79 +1,14 @@
 import * as zarr from "zarrita";
 
+import { decodeVariableChunkInPlace } from "./variableDecoding.ts";
 import { ZarrDataManager } from "./ZarrDataManager.ts";
 
 import { type TSources } from "@/lib/types/GlobeTypes.ts";
 
-export type TDataBounds = {
-  min: number;
-  max: number;
-  fillValue: number;
-  missingValue: number;
+type TLatLonNames = {
+  latitudeName: string | null;
+  longitudeName: string | null;
 };
-
-export function getMissingValue(
-  datavar: zarr.Array<zarr.DataType, zarr.AsyncReadable>
-) {
-  const attributes = datavar.attrs;
-  if (Object.hasOwn(attributes, "missingValue")) {
-    return new Float32Array([Number(attributes.missingValue)])[0];
-  }
-  if (Object.hasOwn(attributes, "missing_value")) {
-    return new Float32Array([Number(attributes.missing_value)])[0];
-  }
-  return NaN;
-}
-
-/**
- * Retrieves the fill value from a Zarr array, normalizing across different
- * naming conventions ("fillValue", "fill_value", "_FillValue", "_fillvalue")
- * that various tools and conventions (Zarr v2, CF conventions, xarray, etc.)
- * may use to store it in the metadata attributes.
- */
-export function getFillValue(
-  datavar: zarr.Array<zarr.DataType, zarr.AsyncReadable>
-) {
-  if (datavar.fillValue) {
-    return datavar.fillValue as number;
-  }
-  const attributes = datavar.attrs;
-  if (Object.hasOwn(attributes, "fillValue")) {
-    return new Float32Array([Number(attributes.fillValue)])[0];
-  }
-  if (Object.hasOwn(attributes, "fill_value")) {
-    return new Float32Array([Number(attributes.fill_value)])[0];
-  }
-  if (Object.hasOwn(attributes, "_FillValue")) {
-    return new Float32Array([Number(attributes._FillValue)])[0];
-  }
-  if (Object.hasOwn(attributes, "_fillvalue")) {
-    return new Float32Array([Number(attributes._fillvalue)])[0];
-  }
-  return NaN;
-}
-
-/**
- * Create a predicate that returns true when a value equals the dataset's
- * missing or fill value (or is NaN).
- */
-export function createMissingOrFillPredicate(
-  datavar: zarr.Array<zarr.DataType, zarr.AsyncReadable>
-) {
-  const missingValue = getMissingValue(datavar);
-  const fillValue = getFillValue(datavar);
-  return (value: number) => {
-    if (Number.isNaN(value)) {
-      return true;
-    }
-    if (value === missingValue) {
-      return true;
-    }
-    if (value === fillValue) {
-      return true;
-    }
-    return false;
-  };
-}
 
 export function hasUnits(
   maybeHasUnits: unknown
@@ -105,6 +40,7 @@ export function isLatitudeVariable(name: string, attrs: unknown) {
     name === "latitude"
   );
 }
+
 export function isLatitudeName(name: string) {
   // FIXME: Need to check for unit later
   // having "rlat" here is a workaround to catch rotated regular grids if the have no CRS-var
@@ -140,7 +76,7 @@ function latPriority(name: string) {
 function resolveLatLonFromCoordinates(
   datavar: zarr.Array<zarr.DataType, zarr.AsyncReadable>,
   isRotated: boolean
-): { latitudeName: string | null; longitudeName: string | null } {
+): TLatLonNames {
   const coordinates = isRotated
     ? "rlon rlat"
     : (datavar.attrs.coordinates as string);
@@ -170,7 +106,7 @@ function refineLatLonFromSources(
   sources: TSources["levels"][0]["datasources"],
   latitudeName: string | null,
   longitudeName: string | null
-): { latitudeName: string | null; longitudeName: string | null } {
+): TLatLonNames {
   for (const sourceKey in sources) {
     if (isLatitudeVariable(sourceKey, sources[sourceKey].attrs)) {
       if (
@@ -230,16 +166,6 @@ function findLatLonNames(
     latitudeName: latitudeName ?? "lat",
     longitudeName: longitudeName ?? "lon",
   };
-}
-
-function applyScaleFactor(
-  data?: zarr.Chunk<zarr.DataType>,
-  attributes?: zarr.Attributes
-) {
-  if (data && attributes?.scale_factor) {
-    const scaleFactor = Number(attributes.scale_factor);
-    data.data = (data.data as Float32Array).map((v) => v * scaleFactor);
-  }
 }
 
 async function fetchLatLonVariables(
@@ -305,88 +231,14 @@ export async function getLatLonData(
     longitudes,
   };
 
-  applyScaleFactor(returnObject.latitudes, returnObject.latitudesAttrs);
-  applyScaleFactor(
+  decodeVariableChunkInPlace(
+    returnObject.latitudes,
+    returnObject.latitudesAttrs
+  );
+  decodeVariableChunkInPlace(
     returnObject.longitudes ?? undefined,
     returnObject.longitudesAttrs ?? undefined
   );
 
   return returnObject;
-}
-
-export function getDataBoundsAndMapMissingToNaN(
-  datavar: zarr.Array<zarr.DataType, zarr.AsyncReadable>,
-  data: Float32Array<ArrayBufferLike>,
-  missingValue = getMissingValue(datavar),
-  fillValue = getFillValue(datavar)
-): TDataBounds {
-  let min = Number.POSITIVE_INFINITY;
-  let max = Number.NEGATIVE_INFINITY;
-
-  for (let i = 0; i < data.length; i++) {
-    const v = data[i];
-    if (v === missingValue || v === fillValue || !Number.isFinite(v)) {
-      data[i] = NaN;
-      continue;
-    }
-    if (v < min) {
-      min = v;
-    }
-    if (v > max) {
-      max = v;
-    }
-  }
-
-  if (min === Number.POSITIVE_INFINITY) {
-    min = NaN;
-  }
-  if (max === Number.NEGATIVE_INFINITY) {
-    max = NaN;
-  }
-  return { min, max, fillValue, missingValue };
-}
-
-export function mapMissingAndFillToNaN(
-  data: Float32Array<ArrayBufferLike>,
-  missingValue: number,
-  fillValue: number
-) {
-  for (let i = 0; i < data.length; i++) {
-    const value = data[i];
-    if (
-      !Number.isFinite(value) ||
-      value === missingValue ||
-      value === fillValue
-    ) {
-      data[i] = NaN;
-    }
-  }
-  return data;
-}
-
-/**
- * Gridlook cannot handle Float64 and integer types in textures, so cast to Float32
- */
-export function castDataVarToFloat32(
-  rawData:
-    | unknown[]
-    | Int8Array<ArrayBufferLike>
-    | Int16Array<ArrayBufferLike>
-    | Int32Array<ArrayBufferLike>
-    | BigInt64Array<ArrayBufferLike>
-    | Uint8Array<ArrayBufferLike>
-    | Uint16Array<ArrayBufferLike>
-    | Uint32Array<ArrayBufferLike>
-    | BigUint64Array<ArrayBufferLike>
-    | Float32Array<ArrayBufferLike>
-    | Float64Array<ArrayBufferLike>
-    | zarr.BoolArray
-    | zarr.UnicodeStringArray
-    | zarr.ByteStringArray
-    | zarr.Chunk<zarr.DataType>
-) {
-  if (rawData instanceof Float32Array) {
-    return rawData;
-  }
-  return Float32Array.from(rawData as ArrayLike<number>);
 }
